git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/branches/networking@13064 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/addons/addons_manager.cpp
@@ -27,14 +27,15 @@
 #include <string.h>
 #include <vector>
 
-#include "addons/network_http.hpp"
+#include "addons/inetwork_http.hpp"
 #include "addons/request.hpp"
 #include "addons/zip.hpp"
-#include "graphics/irr_driver.hpp"
 #include "io/file_manager.hpp"
 #include "io/xml_node.hpp"
+#include "karts/kart_properties.hpp"
 #include "karts/kart_properties_manager.hpp"
 #include "states_screens/kart_selection.hpp"
+#include "tracks/track.hpp"
 #include "tracks/track_manager.hpp"
 #include "utils/string_utils.hpp"
 
@@ -49,14 +50,30 @@ AddonsManager::AddonsManager() : m_addons_list(std::vector<Addon>() ),
                                  m_state(STATE_INIT)
 {
     m_file_installed = file_manager->getAddonsFile("addons_installed.xml");
+
+    // Load the addons list (even if internet is disabled)
+    m_addons_list.lock();
+    // Clear the list in case that a reinit is being done.
+    m_addons_list.getData().clear();
+    loadInstalledAddons();
+    m_addons_list.unlock();
 }   // AddonsManager
+
+// ----------------------------------------------------------------------------
+/** The destructor saves the installed addons file again. This is necessary
+ *  so that information about downloaded icons is saved for the next run.
+ */
+AddonsManager::~AddonsManager()
+{
+    saveInstalled();
+}   // ~AddonsManager
 
 // ----------------------------------------------------------------------------
 /** This initialises the online portion of the addons manager. It uses the
  *  downloaded list of available addons. This is called by network_http before
  *  it goes into command-receiving mode, so we can't use any asynchronous calls
  *  here (though this is being called from a separate thread , so the
- *  main GUI is not blocked anyway). This function will update the state 
+ *  main GUI is not blocked anyway). This function will update the state
  *  variable
  */
 void AddonsManager::initOnline(const XMLNode *xml)
@@ -92,19 +109,23 @@ void AddonsManager::initOnline(const XMLNode *xml)
             else
                 wrong_version = stk_version <stk_config->m_min_track_version ||
                                 stk_version >stk_config->m_max_track_version   ;
+            // If the add-on is included, behave like it is a wrong version
+            if (addon.testIncluded(addon.getMinIncludeVer(), addon.getMaxIncludeVer()))
+                wrong_version = true;
+
             // Check which version to use: only for this stk version,
             // and not addons that are marked as hidden (testing=0)
             if(wrong_version|| testing==0)
             {
                 // If the version is too old (e.g. after an update of stk)
                 // remove a cached icon.
-                std::string full_path = 
+                std::string full_path =
                     file_manager->getAddonsFile("icons/"
                                                 +addon.getIconBasename());
                 if(file_manager->fileExists(full_path))
                 {
                     if(UserConfigParams::logAddons())
-                        printf("[addons] Removing cached icon '%s'.\n", 
+                        Log::warn("[addons] Removing cached icon '%s'.\n",
                                addon.getIconBasename().c_str());
                     file_manager->removeFile(full_path);
                 }
@@ -114,21 +135,28 @@ void AddonsManager::initOnline(const XMLNode *xml)
             m_addons_list.lock();
             if(index>=0)
             {
-                // Only copy the data if a newer revision is found.
-                if(m_addons_list.getData()[index].getRevision()
-                    < addon.getRevision()                       )
+                Addon& tmplist_addon = m_addons_list.getData()[index];
+
+                // Only copy the data if a newer revision is found (ignore unapproved
+                // revisions unless player is in the mode to see them)
+                if (tmplist_addon.getRevision() < addon.getRevision() &&
+                    (addon.testStatus(Addon::AS_APPROVED) || UserConfigParams::m_artist_debug_mode))
+                {
                     m_addons_list.getData()[index].copyInstallData(addon);
+                }
             }
             else
             {
                 m_addons_list.getData().push_back(addon);
                 index = m_addons_list.getData().size()-1;
             }
+            // Mark that this addon still exists on the server
+            m_addons_list.getData()[index].setStillExists();
             m_addons_list.unlock();
         }
         else
         {
-            fprintf(stderr, 
+            fprintf(stderr,
                     "[addons] Found invalid node '%s' while downloading addons.\n",
                     node->getName().c_str());
             fprintf(stderr, "[addons] Ignored.\n");
@@ -136,9 +164,46 @@ void AddonsManager::initOnline(const XMLNode *xml)
     }   // for i<xml->getNumNodes
     delete xml;
 
+    // Now remove all items from the addons-installed list, that are not
+    // on the server anymore (i.e. not in the addons.xml file), and not
+    // installed. If found, remove the icon cached for this addon.
+    // Note that if (due to a bug) an icon is shared (i.e. same icon on
+    // an addon that's still on the server and an invalid entry in the
+    // addons installed file), it will be re-downloaded later.
+    m_addons_list.lock();
+    unsigned int count = m_addons_list.getData().size();
+
+    for(unsigned int i=0; i<count;)
+    {
+        if(m_addons_list.getData()[i].getStillExists() ||
+            m_addons_list.getData()[i].isInstalled())
+        {
+            i++;
+            continue;
+        }
+        // This addon is not on the server anymore, and not installed. Remove
+        // it from the list.
+        if(UserConfigParams::logAddons())
+            Log::warn(
+                "[addons] Removing '%s' which is not on the server anymore.\n",
+                m_addons_list.getData()[i].getId().c_str() );
+        std::string icon = m_addons_list.getData()[i].getIconBasename();
+        std::string icon_file =file_manager->getAddonsFile("icons/"+icon);
+        if(file_manager->fileExists(icon_file))
+        {
+            file_manager->removeFile(icon_file);
+            // Ignore errors silently.
+        }
+        m_addons_list.getData()[i] = m_addons_list.getData()[count-1];
+        m_addons_list.getData().pop_back();
+        count--;
+    }
+    m_addons_list.unlock();
+
     m_state.setAtomic(STATE_READY);
 
-    downloadIcons();
+    if (UserConfigParams::m_internet_status == INetworkHttp::IPERM_ALLOWED)
+        downloadIcons();
 }   // initOnline
 
 // ----------------------------------------------------------------------------
@@ -147,14 +212,68 @@ void AddonsManager::initOnline(const XMLNode *xml)
  */
 void AddonsManager::reInit()
 {
-    m_addons_list.lock();
-    m_addons_list.getData().clear();
-    m_addons_list.unlock();
     m_state.setAtomic(STATE_INIT);
 }   // reInit
 
 // ----------------------------------------------------------------------------
-/** Download all necessary icons (i.e. icons that are either missing or have 
+/** This function checks if the information in the installed addons file is
+ *  consistent with what is actually available. This avoids e.g. that an
+ *  addon is installed, but not marked here (and therefore shows up as
+ *  not installed in the addons GUI), see bug #455.
+ */
+void AddonsManager::checkInstalledAddons()
+{
+    bool something_was_changed = false;
+
+    // Lock the whole addons list to make sure a consistent view is
+    // written back to disk. The network thread might still be
+    // downloading icons and modify content
+    m_addons_list.lock();
+
+    // First karts
+    // -----------
+    for(unsigned int i=0; i<kart_properties_manager->getNumberOfKarts(); i++)
+    {
+        const KartProperties *kp = kart_properties_manager->getKartById(i);
+        const std::string &dir=kp->getKartDir();
+        if(dir.find(file_manager->getAddonsDir())==std::string::npos)
+            continue;
+        int n = getAddonIndex(kp->getIdent());
+        if(n<0) continue;
+        if(!m_addons_list.getData()[n].isInstalled())
+        {
+            Log::info("[addons] Marking '%s' as being installed.\n",
+                   kp->getIdent().c_str());
+            m_addons_list.getData()[n].setInstalled(true);
+            something_was_changed = true;
+        }
+    }
+
+    // Then tracks
+    // -----------
+    for(unsigned int i=0; i<track_manager->getNumberOfTracks(); i++)
+    {
+        const Track *track = track_manager->getTrack(i);
+        const std::string &dir=track->getFilename();
+        if(dir.find(file_manager->getAddonsDir())==std::string::npos)
+            continue;
+        int n = getAddonIndex(track->getIdent());
+        if(n<0) continue;
+        if(!m_addons_list.getData()[n].isInstalled())
+        {
+            Log::info("[addons] Marking '%s' as being installed.\n",
+                   track->getIdent().c_str());
+            m_addons_list.getData()[n].setInstalled(true);
+            something_was_changed = true;
+        }
+    }
+    if(something_was_changed)
+        saveInstalled();
+    m_addons_list.unlock();
+}   // checkInstalledAddons
+
+// ----------------------------------------------------------------------------
+/** Download all necessary icons (i.e. icons that are either missing or have
  *  been updated since they were downloaded).
  */
 void AddonsManager::downloadIcons()
@@ -164,7 +283,7 @@ void AddonsManager::downloadIcons()
         Addon &addon            = m_addons_list.getData()[i];
         const std::string &icon = addon.getIconBasename();
         const std::string &icon_full
-                                = file_manager->getAddonsFile("icons/"+icon); 
+                                = file_manager->getAddonsFile("icons/"+icon);
         if(addon.iconNeedsUpdate() ||
             !file_manager->fileExists(icon_full))
         {
@@ -179,10 +298,11 @@ void AddonsManager::downloadIcons()
                 continue;
             }
             std::string save        = "icons/"+icon;
-            Request *r = network_http->downloadFileAsynchron(url, save, 
-                                                 /*priority*/1,
-                                               /*manage_mem*/false);
-            r->setAddonIconNotification(&addon);            
+            Request *r = INetworkHttp::get()->downloadFileAsynchron(url, save,
+                                                            /*priority*/1,
+                                                           /*manage_mem*/true);
+            if (r != NULL)
+                r->setAddonIconNotification(&addon);
         }
         else
             m_addons_list.getData()[i].setIconReady();
@@ -197,8 +317,11 @@ void AddonsManager::downloadIcons()
 void AddonsManager::loadInstalledAddons()
 {
     /* checking for installed addons */
-    std::cout << "[addons] Loading an xml file for installed addons: ";
-    std::cout << m_file_installed << std::endl;
+    if(UserConfigParams::logAddons())
+    {
+        std::cout << "[addons] Loading an xml file for installed addons: ";
+        std::cout << m_file_installed << std::endl;
+    }
     const XMLNode *xml = file_manager->createXMLTree(m_file_installed);
     if(!xml)
         return;
@@ -218,7 +341,7 @@ void AddonsManager::loadInstalledAddons()
 }   // loadInstalledAddons
 
 // ----------------------------------------------------------------------------
-/** Returns an addon with a given id. Raises an assertion if the id is not 
+/** Returns an addon with a given id. Raises an assertion if the id is not
  *  found!
  *  \param id The id to search for.
  */
@@ -229,6 +352,10 @@ const Addon* AddonsManager::getAddon(const std::string &id) const
 }   // getAddon
 
 // ----------------------------------------------------------------------------
+/** Returns the index of the addon with the given id, or -1 if no such
+ *  addon exist.
+ *  \param id The (unique) identifier of the addon.
+ */
 int AddonsManager::getAddonIndex(const std::string &id) const
 {
     for(unsigned int i = 0; i < m_addons_list.getData().size(); i++)
@@ -242,8 +369,8 @@ int AddonsManager::getAddonIndex(const std::string &id) const
 }   // getAddonIndex
 
 // ----------------------------------------------------------------------------
-/** Installs or updates (i.e. = install on top of an existing installation) an 
- *  addon. It checks for the directories and then unzips the file (which must 
+/** Installs or updates (i.e. = install on top of an existing installation) an
+ *  addon. It checks for the directories and then unzips the file (which must
  *  already have been downloaded).
  *  \param addon Addon data for the addon to install.
  *  \return true if installation was successful.
@@ -251,45 +378,64 @@ int AddonsManager::getAddonIndex(const std::string &id) const
 bool AddonsManager::install(const Addon &addon)
 {
     bool success=true;
-    const std::string &id = addon.getId();
-    file_manager->checkAndCreateDirForAddons(id, addon.getTypeDirectory());
+    file_manager->checkAndCreateDirForAddons(addon.getDataDir());
 
-    //extract the zip in the addons folder called like the addons name    
+    //extract the zip in the addons folder called like the addons name
     std::string base_name = StringUtils::getBasename(addon.getZipFileName());
     std::string from      = file_manager->getAddonsFile("tmp/"+base_name);
     std::string to        = addon.getDataDir();
-    
+
     success = extract_zip(from, to);
     if (!success)
     {
         // TODO: show a message in the interface
-        std::cerr << "[Addons] Failed to unzip '" << from << "' to '" 
+        std::cerr << "[addons] Failed to unzip '" << from << "' to '"
                   << to << "'\n";
+        std::cerr << "[addons] Zip file will not be removed.\n";
         return false;
+    }
+
+    if(!file_manager->removeFile(from))
+    {
+        std::cerr << "[addons] Problems removing temporary file '"
+                  << from << "'.\n";
     }
 
     int index = getAddonIndex(addon.getId());
     assert(index>=0 && index < (int)m_addons_list.getData().size());
     m_addons_list.getData()[index].setInstalled(true);
-    
+
     if(addon.getType()=="kart")
     {
         // We have to remove the mesh of the kart since otherwise it remains
-        // cashed (if a kart is updated), and will therefore be found again 
-        // when reloading the karts. This is important on one hand since we 
+        // cashed (if a kart is updated), and will therefore be found again
+        // when reloading the karts. This is important on one hand since we
         // reload all karts (this function is easily available) and existing
         // karts will not reload their meshes.
-        const KartProperties *prop = 
+        const KartProperties *prop =
             kart_properties_manager->getKart(addon.getId());
-        // If the model already exist (i.e. it's an update, not a new install)
-        // make sure to remove the cached copy of the mesh
+        // If the model already exist, first remove the old kart
         if(prop)
+            kart_properties_manager->removeKart(addon.getId());
+        kart_properties_manager->loadKart(addon.getDataDir());
+    }
+    else if (addon.getType()=="track" || addon.getType()=="arena")
+    {
+        Track *track = track_manager->getTrack(addon.getId());
+        if(track)
+            track_manager->removeTrack(addon.getId());
+
+        try
         {
-            const KartModel &model = prop->getMasterKartModel();
-            irr_driver->removeMeshFromCache(model.getModel());
+            track_manager->loadTrack(addon.getDataDir());
+        }
+        catch (std::exception& e)
+        {
+            fprintf(stderr, "[AddonsManager] ERROR: Cannot load track <%s> : %s\n",
+                    addon.getDataDir().c_str(), e.what());
         }
     }
-    saveInstalled(addon.getType());
+    saveInstalled();
     return true;
 }   // install
 
@@ -300,8 +446,8 @@ bool AddonsManager::install(const Addon &addon)
  */
 bool AddonsManager::uninstall(const Addon &addon)
 {
-    std::cout << "[Addons] Uninstalling <" 
-              << addon.getName() << ">\n";
+    std::cout << "[addons] Uninstalling <"
+              << core::stringc(addon.getName()).c_str() << ">\n";
 
     // addon is a const reference, and to avoid removing the const, we
     // find the proper index again to modify the installed state
@@ -310,13 +456,33 @@ bool AddonsManager::uninstall(const Addon &addon)
     m_addons_list.getData()[index].setInstalled(false);
 
     //remove the addons directory
-    bool error = !file_manager->removeDirectory(addon.getDataDir());
-    saveInstalled(addon.getType());
+	bool error = false;
+	// if the user deleted the data directory for an add-on with
+	// filesystem tools, removeTrack/removeKart will trigger an assert
+	// because the kart/track was never added in the first place
+	if (file_manager->fileExists(addon.getDataDir()))
+	{
+		error = !file_manager->removeDirectory(addon.getDataDir());
+		if(addon.getType()=="kart")
+		{
+			kart_properties_manager->removeKart(addon.getId());
+		}
+		else if(addon.getType()=="track" || addon.getType()=="arena")
+		{
+			track_manager->removeTrack(addon.getId());
+		}
+	}
+    saveInstalled();
     return !error;
 }   // uninstall
 
 // ----------------------------------------------------------------------------
-void AddonsManager::saveInstalled(const std::string &type)
+/** Saves the information about installed addons and cached icons to
+ *  addons_installed.xml. If this is not called, information about downloaded
+ *  icons is lost (and will trigger a complete redownload when STK is started
+ *  next time).
+ */
+void AddonsManager::saveInstalled()
 {
     //Put the addons in the xml file
     //Manually because the irrlicht xml writer doesn't seem finished, FIXME ?
@@ -333,9 +499,5 @@ void AddonsManager::saveInstalled(const std::string &type)
     }
     xml_installed << "</addons>" << std::endl;
     xml_installed.close();
-    if(type=="kart")
-        kart_properties_manager->reLoadAllKarts();
-    else if(type=="track" || type=="arena")
-        track_manager->loadTrackList();
 }   // saveInstalled
